Move reports files to stage folder
