+++ cypress/e2e/stage/reports/HostAnomaliesBar.spec.js
@@ -1,4 +1,4 @@
-import { hab as pageName } from '../../support/constants';
+import { hab as pageName } from '../../../support/constants';
 //disable as there's no data for this page on eph. env. right now
 xdescribe('Report: Slow hosts by job template (aka Host Anomalies Bar)', () => {
   beforeEach(() => {
